Move automapper out of the  DI framework and put crude detection to prevent it from initializing more than once.
